move schemas to /schema folder
